requirements.txt
PyAudio==0.2.14
google-cloud-texttospeech==2.26.0
google-cloud-speech==2.24.1
anthropic==0.50.0
python-dotenv==1.0.1
httpx==0.28.1
google-api-core==2.25.0rc0
google-auth==2.40.0
protobuf==4.25.7
grpcio==1.71.0
grpcio-status==1.62.3
urllib3==1.26.15
flask
jamo
numpy
requests 